+++ src/Settings/ModuleSettings.php
@@ -14,7 +14,6 @@ use OxidEsales\EshopCommunity\Internal\Framework\Module\Facade\ModuleSettingServ
 
 class ModuleSettings implements ModuleSettingsInterface
 {
-    public const SETTING_DOCUMENT_FOOTER = 'fa_invoice_DocumentFooter';
     public const SETTING_DOCUMENT_FILENAME_PREFIX = 'fa_invoice_FilenamePrefix';
     public const SETTING_DOCUMENT_IS_FOR_ARCHIVE = 'fa_invoice_IsForArchive';
     public const SETTING_INVOICE_NUMBER_FORMAT = 'fa_invoice_InvoiceNumberFormat';
@@ -22,11 +21,6 @@ class ModuleSettings implements ModuleSettingsInterface
     public function __construct(
         private ModuleSettingServiceInterface $moduleSettingService
     ) {
-    }
-
-    public function getDocumentFooter(): string
-    {
-        return $this->getStringSetting(self::SETTING_DOCUMENT_FOOTER);
     }
 
     public function getFilePrefix(): string