+++ src/Settings/ModuleSettingsInterface.php
@@ -11,8 +11,6 @@ namespace FreshAdvance\Invoice\Settings;
 
 interface ModuleSettingsInterface
 {
-    public function getDocumentFooter(): string;
-
     public function getFilePrefix(): string;
 
     public function isForArchive(): bool;